+++ app/src/Form/Field/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ErrorInterface, FormContextInterface, FormContext, ValuesInterface } from '.';
+import { ErrorInterface, FormContextInterface, FormContext, ValuesInterface } from '..';
 
 type Editor = "textbox" | "multilinetextbox" | "dropdown" | "email" | "password";
